+++ apps/web/app/future/[user]/[type]/page.tsx
@@ -4,10 +4,10 @@ import { WithLayout } from "app/layoutHOC";
 import { type GetServerSidePropsContext } from "next";
 import { headers, cookies } from "next/headers";
 
-import { getServerSideProps, type PageProps } from "@lib/[user]/[type]/getServerSideProps";
 import { buildLegacyCtx } from "@lib/buildLegacyCtx";
 
-import LegacyPage from "@components/pages/[user]/[type]";
+import LegacyPage from "~/users/views/users-type-public-view";
+import { getServerSideProps, type PageProps } from "~/users/views/users-type-public-view.getServerSideProps";
 
 export const generateMetadata = async ({
   params,
