+++ apps/web/modules/users/views/users-type-public-view.tsx
@@ -6,7 +6,7 @@ import { Booker } from "@calcom/atoms";
 import { getBookerWrapperClasses } from "@calcom/features/bookings/Booker/utils/getBookerWrapperClasses";
 import { BookerSeo } from "@calcom/features/bookings/components/BookerSeo";
 
-import { type PageProps } from "@lib/[user]/[type]/getServerSideProps";
+import { type PageProps } from "./users-type-public-view.getServerSideProps";
 
 export const getMultipleDurationValue = (
   multipleDurationConfig: number[] | undefined,
