+++ apps/web/pages/[user]/[type]/index.tsx
@@ -1,10 +1,10 @@
-import { getServerSideProps, type PageProps } from "@lib/[user]/[type]/getServerSideProps";
 import { type AppProps } from "@lib/app-providers";
 
 import PageWrapper, { type CalPageWrapper } from "@components/PageWrapper";
-import TypePage from "@components/pages/[user]/[type]";
 
-export { getServerSideProps, PageProps };
+import TypePage from "~/users/views/users-type-public-view";
+
+export { getServerSideProps } from "~/users/views/users-type-public-view.getServerSideProps";
 
 const Type = TypePage as unknown as CalPageWrapper & {
   isBookingPage: AppProps["Component"]["isBookingPage"];
