+++ apps/web/pages/org/[orgSlug]/[user]/[type]/index.tsx
@@ -1,11 +1,12 @@
 "use client";
 
 import PageWrapper from "@components/PageWrapper";
-import UserTypePage from "@components/pages/[user]/[type]";
 
-import type { PageProps as UserTypePageProps } from "../../../[user]/[type]";
-import type { PageProps as TeamTypePageProps } from "../../../team/[slug]/[type]";
-import TeamTypePage from "../../../team/[slug]/[type]";
+import UserTypePage from "~/users/views/users-type-public-view";
+import type { PageProps as UserTypePageProps } from "~/users/views/users-type-public-view.getServerSideProps";
+
+import type { PageProps as TeamTypePageProps } from "../../../../team/[slug]/[type]";
+import TeamTypePage from "../../../../team/[slug]/[type]";
 
 export { getServerSideProps } from "@lib/org/[orgSlug]/[user]/[type]/getServerSideProps";
